Remove TracingInterface
This was initially introduced to be implemented both in `LinuxTracing` and
`WindowsTracing`, but then we decided to go a different way, so let's remove it
as discussed.
This effectively reverts commit 7c9ef2809490533b43fa395c9af74ec13318a762.

Test: Capture Trata.
