+++ CMakeLists.txt
@@ -187,7 +187,6 @@ add_subdirectory(src/OrbitVersion)
 add_subdirectory(src/PresetFile)
 add_subdirectory(src/StringManager)
 add_subdirectory(src/Symbols)
-add_subdirectory(src/TracingInterface)
 add_subdirectory(src/Test)
 add_subdirectory(src/TestUtils)
