+++ src/LinuxTracing/include/LinuxTracing/Tracer.h
@@ -10,15 +10,15 @@
 #include <thread>
 #include <utility>
 
-#include "TracingInterface/Tracer.h"
+#include "LinuxTracing/TracerListener.h"
 #include "capture.pb.h"
 
 namespace orbit_linux_tracing {
 
-class Tracer : public orbit_tracing_interface::Tracer {
+class Tracer {
  public:
   explicit Tracer(orbit_grpc_protos::CaptureOptions capture_options)
-      : orbit_tracing_interface::Tracer(std::move(capture_options)) {}
+      : capture_options_{std::move(capture_options)} {}
 
   ~Tracer() { Stop(); }
 
@@ -27,12 +27,14 @@ class Tracer : public orbit_tracing_interface::Tracer {
   Tracer(Tracer&&) = default;
   Tracer& operator=(Tracer&&) = default;
 
-  void Start() override {
+  void SetListener(TracerListener* listener) { listener_ = listener; }
+
+  void Start() {
     *exit_requested_ = false;
     thread_ = std::make_shared<std::thread>(&Tracer::Run, this);
   }
 
-  void Stop() override {
+  void Stop() {
     *exit_requested_ = true;
     if (thread_ != nullptr && thread_->joinable()) {
       thread_->join();
@@ -41,6 +43,10 @@ class Tracer : public orbit_tracing_interface::Tracer {
   }
 
  private:
+  orbit_grpc_protos::CaptureOptions capture_options_;
+
+  TracerListener* listener_ = nullptr;
+
   // exit_requested_ must outlive this object because it is used by thread_.
   // The control block of shared_ptr is thread safe (i.e., reference counting
   // and pointee's lifetime management are atomic and thread safe).
