+++ src/LinuxTracing/include/LinuxTracing/TracerListener.h
@@ -2,12 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef TRACING_TRACER_LISTENER_H_
-#define TRACING_TRACER_LISTENER_H_
+#ifndef LINUX_TRACING_TRACER_LISTENER_H_
+#define LINUX_TRACING_TRACER_LISTENER_H_
 
 #include "capture.pb.h"
 
-namespace orbit_tracing_interface {
+namespace orbit_linux_tracing {
 
 class TracerListener {
  public:
@@ -32,6 +32,6 @@ class TracerListener {
       orbit_grpc_protos::OutOfOrderEventsDiscardedEvent out_of_order_events_discarded_event) = 0;
 };
 
-}  // namespace orbit_tracing_interface
+}  // namespace orbit_linux_tracing
 
-#endif  // TRACING_TRACER_LISTENER_H_
+#endif  // LINUX_TRACING_TRACER_LISTENER_H_
