+++ src/LinuxTracingIntegrationTests/LinuxTracingIntegrationTest.cpp
@@ -19,6 +19,7 @@
 #include <utility>
 
 #include "LinuxTracing/Tracer.h"
+#include "LinuxTracing/TracerListener.h"
 #include "LinuxTracingIntegrationTestPuppet.h"
 #include "ObjectUtils/Address.h"
 #include "ObjectUtils/ElfFile.h"
@@ -27,7 +28,6 @@
 #include "OrbitBase/Logging.h"
 #include "OrbitBase/ReadFileToString.h"
 #include "OrbitBase/ThreadUtils.h"
-#include "TracingInterface/TracerListener.h"
 #include "capture.pb.h"
 
 namespace orbit_linux_tracing_integration_tests {
@@ -162,7 +162,7 @@ class ChildProcess {
   int writing_fd_ = -1;
 };
 
-class BufferTracerListener : public orbit_tracing_interface::TracerListener {
+class BufferTracerListener : public orbit_linux_tracing::TracerListener {
  public:
   void OnSchedulingSlice(orbit_grpc_protos::SchedulingSlice scheduling_slice) override {
     orbit_grpc_protos::ProducerCaptureEvent event;
